dash-board/python-booking-app/requirements.txt
Flask==2.1.2
mysql-connector-python==8.0.28
pandas==1.4.2
numpy==1.22.3
WTForms==3.0.1